+++ fdbclient/GenericManagementAPI.actor.h
@@ -129,21 +129,6 @@ bool isCompleteConfiguration(std::map<std::string, std::string> const& options);
 
 ConfigureAutoResult parseConfig(StatusObject const& status);
 
-template <typename Transaction, class T>
-struct transaction_future_type {
-	using type = typename Transaction::template FutureT<T>;
-};
-
-template <typename Transaction, class T>
-struct transaction_future_type<Transaction*, T> {
-	using type = typename transaction_future_type<Transaction, T>::type;
-};
-
-template <typename Transaction, class T>
-struct transaction_future_type<Reference<Transaction>, T> {
-	using type = typename transaction_future_type<Transaction, T>::type;
-};
-
 // Management API written in template code to support both IClientAPI and NativeAPI
 namespace ManagementAPI {
 
@@ -666,258 +651,6 @@ Future<ConfigurationResult> changeConfig(Reference<DB> db,
 // used by special keys and fdbcli
 std::string generateErrorMessage(const CoordinatorsResult& res);
 
-ACTOR template <class Transaction>
-Future<Optional<TenantMapEntry>> tryGetTenantTransaction(Transaction tr, TenantName name) {
-	state Key tenantMapKey = name.withPrefix(tenantMapPrefix);
-
-	tr->setOption(FDBTransactionOptions::RAW_ACCESS);
-
-	state typename transaction_future_type<Transaction, Optional<Value>>::type tenantFuture = tr->get(tenantMapKey);
-	Optional<Value> val = wait(safeThreadFutureToFuture(tenantFuture));
-	return val.map<TenantMapEntry>([](Optional<Value> v) { return decodeTenantEntry(v.get()); });
-}
-
-ACTOR template <class DB>
-Future<Optional<TenantMapEntry>> tryGetTenant(Reference<DB> db, TenantName name) {
-	state Reference<typename DB::TransactionT> tr = db->createTransaction();
-
-	loop {
-		try {
-			tr->setOption(FDBTransactionOptions::READ_SYSTEM_KEYS);
-			tr->setOption(FDBTransactionOptions::READ_LOCK_AWARE);
-			Optional<TenantMapEntry> entry = wait(tryGetTenantTransaction(tr, name));
-			return entry;
-		} catch (Error& e) {
-			wait(safeThreadFutureToFuture(tr->onError(e)));
-		}
-	}
-}
-
-ACTOR template <class Transaction>
-Future<TenantMapEntry> getTenantTransaction(Transaction tr, TenantName name) {
-	Optional<TenantMapEntry> entry = wait(tryGetTenantTransaction(tr, name));
-	if (!entry.present()) {
-		throw tenant_not_found();
-	}
-
-	return entry.get();
-}
-
-ACTOR template <class DB>
-Future<TenantMapEntry> getTenant(Reference<DB> db, TenantName name) {
-	Optional<TenantMapEntry> entry = wait(tryGetTenant(db, name));
-	if (!entry.present()) {
-		throw tenant_not_found();
-	}
-
-	return entry.get();
-}
-
-// Creates a tenant with the given name. If the tenant already exists, an empty optional will be returned.
-// The caller must enforce that the tenant ID be unique from all current and past tenants, and it must also be unique
-// from all other tenants created in the same transaction.
-ACTOR template <class Transaction>
-Future<std::pair<TenantMapEntry, bool>> createTenantTransaction(Transaction tr, TenantNameRef name, int64_t tenantId) {
-	state Key tenantMapKey = name.withPrefix(tenantMapPrefix);
-
-	if (name.startsWith("\xff"_sr)) {
-		throw invalid_tenant_name();
-	}
-
-	tr->setOption(FDBTransactionOptions::RAW_ACCESS);
-
-	state Future<Optional<TenantMapEntry>> tenantEntryFuture = tryGetTenantTransaction(tr, name);
-	state typename transaction_future_type<Transaction, Optional<Value>>::type tenantDataPrefixFuture =
-	    tr->get(tenantDataPrefixKey);
-	state typename transaction_future_type<Transaction, Optional<Value>>::type tenantModeFuture =
-	    tr->get(configKeysPrefix.withSuffix("tenant_mode"_sr));
-
-	Optional<Value> tenantMode = wait(safeThreadFutureToFuture(tenantModeFuture));
-
-	if (!tenantMode.present() || tenantMode.get() == StringRef(format("%d", TenantMode::DISABLED))) {
-		throw tenants_disabled();
-	}
-
-	Optional<TenantMapEntry> tenantEntry = wait(tenantEntryFuture);
-	if (tenantEntry.present()) {
-		return std::make_pair(tenantEntry.get(), false);
-	}
-
-	Optional<Value> tenantDataPrefix = wait(safeThreadFutureToFuture(tenantDataPrefixFuture));
-	if (tenantDataPrefix.present() &&
-	    tenantDataPrefix.get().size() + TenantMapEntry::ROOT_PREFIX_SIZE > CLIENT_KNOBS->TENANT_PREFIX_SIZE_LIMIT) {
-		TraceEvent(SevWarnAlways, "TenantPrefixTooLarge")
-		    .detail("TenantSubspace", tenantDataPrefix.get())
-		    .detail("TenantSubspaceLength", tenantDataPrefix.get().size())
-		    .detail("RootPrefixLength", TenantMapEntry::ROOT_PREFIX_SIZE)
-		    .detail("MaxTenantPrefixSize", CLIENT_KNOBS->TENANT_PREFIX_SIZE_LIMIT);
-
-		throw client_invalid_operation();
-	}
-
-	state TenantMapEntry newTenant(tenantId, tenantDataPrefix.present() ? (KeyRef)tenantDataPrefix.get() : ""_sr);
-
-	state typename transaction_future_type<Transaction, RangeResult>::type prefixRangeFuture =
-	    tr->getRange(prefixRange(newTenant.prefix), 1);
-	RangeResult contents = wait(safeThreadFutureToFuture(prefixRangeFuture));
-	if (!contents.empty()) {
-		throw tenant_prefix_allocator_conflict();
-	}
-
-	tr->set(tenantMapKey, encodeTenantEntry(newTenant));
-
-	return std::make_pair(newTenant, true);
-}
-
-ACTOR template <class DB>
-Future<TenantMapEntry> createTenant(Reference<DB> db, TenantName name) {
-	state Reference<typename DB::TransactionT> tr = db->createTransaction();
-
-	state bool firstTry = true;
-	loop {
-		try {
-			tr->setOption(FDBTransactionOptions::ACCESS_SYSTEM_KEYS);
-			tr->setOption(FDBTransactionOptions::LOCK_AWARE);
-			state typename DB::TransactionT::template FutureT<Optional<Value>> lastIdFuture = tr->get(tenantLastIdKey);
-
-			if (firstTry) {
-				Optional<TenantMapEntry> entry = wait(tryGetTenantTransaction(tr, name));
-				if (entry.present()) {
-					throw tenant_already_exists();
-				}
-
-				firstTry = false;
-			}
-
-			Optional<Value> lastIdVal = wait(safeThreadFutureToFuture(lastIdFuture));
-			int64_t tenantId = lastIdVal.present() ? TenantMapEntry::prefixToId(lastIdVal.get()) + 1 : 0;
-			tr->set(tenantLastIdKey, TenantMapEntry::idToPrefix(tenantId));
-			state std::pair<TenantMapEntry, bool> newTenant = wait(createTenantTransaction(tr, name, tenantId));
-
-			if (BUGGIFY) {
-				throw commit_unknown_result();
-			}
-
-			wait(safeThreadFutureToFuture(tr->commit()));
-
-			if (BUGGIFY) {
-				throw commit_unknown_result();
-			}
-
-			TraceEvent("CreatedTenant")
-			    .detail("Tenant", name)
-			    .detail("TenantId", newTenant.first.id)
-			    .detail("Prefix", newTenant.first.prefix)
-			    .detail("Version", tr->getCommittedVersion());
-
-			return newTenant.first;
-		} catch (Error& e) {
-			wait(safeThreadFutureToFuture(tr->onError(e)));
-		}
-	}
-}
-
-ACTOR template <class Transaction>
-Future<Void> deleteTenantTransaction(Transaction tr, TenantNameRef name) {
-	state Key tenantMapKey = name.withPrefix(tenantMapPrefix);
-
-	tr->setOption(FDBTransactionOptions::RAW_ACCESS);
-
-	state Optional<TenantMapEntry> tenantEntry = wait(tryGetTenantTransaction(tr, name));
-	if (!tenantEntry.present()) {
-		return Void();
-	}
-
-	state typename transaction_future_type<Transaction, RangeResult>::type prefixRangeFuture =
-	    tr->getRange(prefixRange(tenantEntry.get().prefix), 1);
-	RangeResult contents = wait(safeThreadFutureToFuture(prefixRangeFuture));
-	if (!contents.empty()) {
-		throw tenant_not_empty();
-	}
-
-	tr->clear(tenantMapKey);
-
-	return Void();
-}
-
-ACTOR template <class DB>
-Future<Void> deleteTenant(Reference<DB> db, TenantName name) {
-	state Reference<typename DB::TransactionT> tr = db->createTransaction();
-
-	state bool firstTry = true;
-	loop {
-		try {
-			tr->setOption(FDBTransactionOptions::ACCESS_SYSTEM_KEYS);
-			tr->setOption(FDBTransactionOptions::LOCK_AWARE);
-
-			if (firstTry) {
-				Optional<TenantMapEntry> entry = wait(tryGetTenantTransaction(tr, name));
-				if (!entry.present()) {
-					throw tenant_not_found();
-				}
-
-				firstTry = false;
-			}
-
-			wait(deleteTenantTransaction(tr, name));
-
-			if (BUGGIFY) {
-				throw commit_unknown_result();
-			}
-
-			wait(safeThreadFutureToFuture(tr->commit()));
-
-			if (BUGGIFY) {
-				throw commit_unknown_result();
-			}
-
-			TraceEvent("DeletedTenant").detail("Tenant", name).detail("Version", tr->getCommittedVersion());
-			return Void();
-		} catch (Error& e) {
-			wait(safeThreadFutureToFuture(tr->onError(e)));
-		}
-	}
-}
-
-ACTOR template <class Transaction>
-Future<std::map<TenantName, TenantMapEntry>> listTenantsTransaction(Transaction tr,
-                                                                    TenantNameRef begin,
-                                                                    TenantNameRef end,
-                                                                    int limit) {
-	state KeyRange range = KeyRangeRef(begin, end).withPrefix(tenantMapPrefix);
-
-	tr->setOption(FDBTransactionOptions::RAW_ACCESS);
-
-	state typename transaction_future_type<Transaction, RangeResult>::type listFuture =
-	    tr->getRange(firstGreaterOrEqual(range.begin), firstGreaterOrEqual(range.end), limit);
-	RangeResult results = wait(safeThreadFutureToFuture(listFuture));
-
-	std::map<TenantName, TenantMapEntry> tenants;
-	for (auto kv : results) {
-		tenants[kv.key.removePrefix(tenantMapPrefix)] = decodeTenantEntry(kv.value);
-	}
-
-	return tenants;
-}
-
-ACTOR template <class DB>
-Future<std::map<TenantName, TenantMapEntry>> listTenants(Reference<DB> db,
-                                                         TenantName begin,
-                                                         TenantName end,
-                                                         int limit) {
-	state Reference<typename DB::TransactionT> tr = db->createTransaction();
-
-	loop {
-		try {
-			tr->setOption(FDBTransactionOptions::READ_SYSTEM_KEYS);
-			tr->setOption(FDBTransactionOptions::READ_LOCK_AWARE);
-			std::map<TenantName, TenantMapEntry> tenants = wait(listTenantsTransaction(tr, begin, end, limit));
-			return tenants;
-		} catch (Error& e) {
-			wait(safeThreadFutureToFuture(tr->onError(e)));
-		}
-	}
-}
 } // namespace ManagementAPI
 
 #include "flow/unactorcompiler.h"
